combine module tests to scraper_test.go
